allocates memory using malloc

+++ 0x0C-more_malloc_free/101-mul.c
@@ -35,7 +35,7 @@ exit(98);
 }
 i++;
 }
-printf("%lld", (long long)argv[1] * (long long)argv[2]);
+printf("%ld", (long)argv[1] * (long)argv[2]);
 return (0);
 }
 }
